first pass of naming changes - plus some auto formatting
